Distinguish instrumentation and sampling for TAU.
Add new option Tool.TAU.mode in config file. The mode should be set as
"instrumentation" if execmd is is already instrumented (with PDT or
compiler). If you want to use sampling, then set the mode as "sampling".
For sampling, execmd could be instrumented or uninstrumented, it doesn't
matter.

Maybe we should give this option a default value.