Add PAPI counter partitioner
The partitioner will try to partition the given counters using a
greedy strategy. The result will be cached into a SQLite database.

The partitioner is implemented as a C++ extension of Python, so it
can be called directly by autoperf. It can also be compiled as a
stand-alone command line program.