Add log facility so the actions can be replayed later
1. New config options:
  - [Logger].loglvl
    Verbosity level of emitted logs. Valid values in increasing
    verbosity order: critical, error, warning, info, verbose,
    command, debug.

    Default is "command".

  - [Logger].logfile
    The file to save the logs.

    Default is <rootdir>/autoperf.log

2. Action replay

When log verbosity level is higher or equal to "command", the log will
be saved as an shell script. Actions taken by autoperf appear as the
shell command and other informations emit appear as the comment. User
can use the log to replay the actions taken by autoperf to run/analyze
the experiment.