Add "hotspots" for sampling data unrolling, plus several bug fixes
A new "hotspots" option for [Analyses.compare2] is implemented.
"hotspots" is a list of pattern strings which will be used to
re.search() on the file name attached to an unrolled method. The
method is a "hotspot" if re.search() returns a match.

"hotspots" is optional. When it is specified, collected data for
a sampling site will be attributed to the first "hotspot" instead
of the first resolved method.